Type is short!

+++ syscalls/java/include/java/io.h
@@ -30,7 +30,7 @@ typedef int NOPH_EOFException_t;
 /* Input stream stuff */
 int NOPH_InputStream_available(NOPH_InputStream_t is); /* Throws */
 int NOPH_InputStream_read(NOPH_InputStream_t is); /* Throws */
-int NOPH_InputStream_read_into(NOPH_InputStream_t is, char* vec, int size, int* eof_ptr); /* Not generated */
+int NOPH_InputStream_read_into(NOPH_InputStream_t is, char* vec, int size, short* eof_ptr); /* Not generated */
 void NOPH_InputStream_reset(NOPH_InputStream_t is); /* Throws */
 void NOPH_InputStream_mark(NOPH_InputStream_t is, int readLimit);
 void NOPH_InputStream_markSupported(NOPH_InputStream_t is); /* Throws */
